Make auto orientation true by default.
